gitignore changes for CLion + MSVC adjustments

+++ bench_OpenMP.cpp
@@ -16,9 +16,9 @@ static void BM_OpenMP(benchmark::State& state) {
   // Code inside this loop is measured repeatedly
   for (auto _ : state) {
     # pragma omp parallel for num_threads(16)
-    for (size_t i = 0; i < 10000000000; ++i) {
+    for (int i = 0; i < 10000000000; ++i) {
       //std::cout << i << '\n';
-      size_t j = i;
+      int j = i;
       benchmark::DoNotOptimize(j);
     }
   }
